Move solution invalid into a helper class at a higher level that also defines the mooseInfo/Warning/Error so we no longer need to do "usingSolutionInvalidWarnings" in every user class refs #28181
