+++ framework/include/actions/Action.h
@@ -15,7 +15,6 @@
 #include "Registry.h"
 #include "PerfGraphInterface.h"
 #include "MooseObjectParameterName.h"
-#include "SolutionInvalidInterface.h"
 
 #include <string>
 #include <ostream>
@@ -31,10 +30,7 @@ class Factory;
 /**
  * Base class for actions.
  */
-class Action : public ParallelParamObject,
-               public MeshMetaDataInterface,
-               public PerfGraphInterface,
-               public SolutionInvalidInterface
+class Action : public ParallelParamObject, public MeshMetaDataInterface, public PerfGraphInterface
 {
 public:
   /// The name of the parameter that contains the unique action name
@@ -50,9 +46,6 @@ public:
    * The method called externally that causes the action to act()
    */
   void timedAct();
-
-  // To get warnings tracked in the SolutionInvalidityOutput
-  usingCombinedWarningSolutionWarnings;
 
 private:
   /**
