+++ framework/src/base/MooseObject.C
@@ -43,9 +43,7 @@ MooseObject::validParams()
 }
 
 MooseObject::MooseObject(const InputParameters & parameters)
-  : ParallelParamObject(parameters),
-    SolutionInvalidInterface(this, parameters),
-    _enabled(getParam<bool>("enable"))
+  : ParallelParamObject(parameters), _enabled(getParam<bool>("enable"))
 {
   if (Registry::isRegisteredObj(type()) && _app.getFactory().currentlyConstructing() != &parameters)
     mooseError(
@@ -62,7 +60,7 @@ const std::string not_shared_error =
 #ifdef MOOSE_KOKKOS_ENABLED
 MooseObject::MooseObject(const MooseObject & object, const Moose::Kokkos::FunctorCopy &)
   : ParallelParamObject(object),
-    SolutionInvalidInterface(this, object.parameters()),
+    MooseVerbosityHelper(this, object.parameters()),
     _enabled(object._enabled)
 {
 }
